+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-07-28
+
+### Added
+- HTML rendering support - Toasts can now render HTML content with the `html` option
+- Configurable animation duration - Animation timing can be customized (defaults to 400ms)
+
+### Fixed
+- DaisyUI style conflict documentation
+
 ## [0.1.0] - 2025-07-18
 
 Initial release